Test: pruebas unitarias elimiar-ingrediente en verde

+++ backend/modelos/ingredientes_servicio.py
@@ -25,7 +25,7 @@ class IngredientesServicio:
         return ingredientes
     
     def agregar_ingrediente(self,ingrediente:Ingrediente) -> Ingrediente:
-        session = Session()
+        session = Session()        
         if(ingrediente.nombre is None or ingrediente.nombre == ""):
             raise ValueError("El nombre del ingrediente es requerido")
         if(ingrediente.tipo is None or ingrediente.tipo == ""):
@@ -33,16 +33,23 @@ class IngredientesServicio:
         if(ingrediente.unidad_medida is None or ingrediente.unidad_medida == ""):
             raise ValueError("La unidad de medida del ingrediente es requerida")
         if(ingrediente.disponible is None or ingrediente.disponible == ""):
-            raise ValueError("La disponibilidad del ingrediente es requerida")
+            raise ValueError("La disponibilidad del ingrediente es requerida")        
         session.add(ingrediente)
         session.commit()
+        session.refresh(ingrediente)
         return ingrediente
+    
 
-    def eliminar_ingrediente(self,ingrediente:Ingrediente) -> Ingrediente:
+    #region para eliminar ingredientes
+    def eliminar_ingrediente(self,ingrediente_id:int) -> (int,str):
         session = Session()
+        ingrediente = session.query(Ingrediente).filter(Ingrediente.id == ingrediente_id).first()
+        if(ingrediente is None):
+            return (404,"El ingrediente no existe")
         tieneReceta = session.query(Receta).filter(Receta.ingrediente_id == ingrediente.id).count() > 0
         if(tieneReceta):
-            raise ValueError("El ingrediente tiene recetas asociadas y no puede ser eliminado")
+            return (400,"El ingrediente tiene recetas asociadas y no puede ser eliminado")
         session.delete(ingrediente)
         session.commit()
-        return ingrediente
+        return (200,"El ingrediente se eliminó correctamente")
+    #endregion
